+++ Platforms/OpenGL3Platform/src/MyGUI_OpenGL3RenderManager.cpp
@@ -57,13 +57,12 @@ namespace MyGUI
 
 	std::string OpenGL3RenderManager::loadFileContent(const std::string& _file)
 	{
-		IDataStream* stream = DataManager::getInstance().getData(_file);
-		if (stream == nullptr)
+		auto stream = DataManager::getInstance().getDataHolder(_file);
+		if (!stream)
 		{
 			MYGUI_PLATFORM_LOG(Error, "Failed to load file content '" << _file << "'.");
 			return {};
 		}
-		DataStreamHolder streamHolder(stream);
 		return stream->readAllText();
 	}